requirements.txt
Flask
jsonify
Werkzeug
python-dotenv
boto3
botocore
pytest